+++ python/GafferSceneTest/EncapsulateTest.py
@@ -298,8 +298,8 @@ class EncapsulateTest( GafferSceneTest.SceneTestCase ) :
 		standardAttributes = GafferScene.StandardAttributes()
 		standardAttributes["in"].setInput( group["out"] )
 		standardAttributes["filter"].setInput( pathFilter["out"] )
-		standardAttributes["attributes"]["transformBlurSegments"]["value"].setValue( 4 )
-		standardAttributes["attributes"]["transformBlurSegments"]["enabled"].setValue( True )
+		standardAttributes["attributes"]["gaffer:transformBlurSegments"]["value"].setValue( 4 )
+		standardAttributes["attributes"]["gaffer:transformBlurSegments"]["enabled"].setValue( True )
 
 		standardOptions = GafferScene.StandardOptions()
 		standardOptions["in"].setInput( standardAttributes["out"] )